Add additional apache configs
According to azure support

Thank you for contacting Microsoft. This is Saif and I’m the Support Engineer who will be working with you on your case # 2112130050002660.

I apologize for the delay in getting back to you. I have reviewed the information and the article that you have provided in your support request.
I understand that you are experiencing an issue with using Storage mounts and file shares on containerized Apache app.

We are suspecting that this issue related to a known behavior for images that use Apache, I will validate this further.
In the meantime, can you please try following the steps below:

    Modify apache configuration to disable mmap and sendfile support. You can find more about these features here:

    http://httpd.apache.org/docs/current/mod/core.html#enablemmap
    http://httpd.apache.org/docs/current/mod/core.html#enablesendfile

Apache configuration snippet:
<Directory "${APACHE_DOCUMENT_ROOT}">
           Options Indexes FollowSymLinks
    EnableMMAP Off
           EnableSendfile Off
    AllowOverride None
    Require all granted
</Directory>